# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Wardan Online Arabic Marketplace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Category {
  id: number;
  slug: string;
  nameAr: string;
  nameEn: string;
  icon: string;
  listingCount: number;
}

export type ListingStatus = typeof ListingStatus[keyof typeof ListingStatus];


export const ListingStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface Listing {
  id: number;
  titleAr: string;
  /** @nullable */
  descriptionAr?: string | null;
  categorySlug: string;
  /** @nullable */
  categoryNameAr?: string | null;
  /** @nullable */
  price: number | null;
  /** @nullable */
  priceUnit?: string | null;
  /** @nullable */
  location?: string | null;
  whatsappNumber: string;
  /** @nullable */
  sellerName?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  status: ListingStatus;
  featured?: boolean;
  createdAt: string;
}

export interface ListingsPage {
  listings: Listing[];
  total: number;
  page: number;
  limit: number;
}

export interface ListingInput {
  /** @minLength 1 */
  titleAr: string;
  descriptionAr?: string;
  categorySlug: string;
  price?: number;
  priceUnit?: string;
  location?: string;
  whatsappNumber: string;
  sellerName?: string;
  imageUrl?: string;
}

export type ListingUpdateStatus = typeof ListingUpdateStatus[keyof typeof ListingUpdateStatus];


export const ListingUpdateStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface ListingUpdate {
  titleAr?: string;
  descriptionAr?: string;
  price?: number;
  priceUnit?: string;
  location?: string;
  whatsappNumber?: string;
  sellerName?: string;
  imageUrl?: string;
  status?: ListingUpdateStatus;
  featured?: boolean;
}

export interface Appointment {
  id: number;
  doctorNameAr: string;
  /** @nullable */
  specialty: string | null;
  /** @nullable */
  clinicLocation?: string | null;
  whatsappNumber: string;
  /** @nullable */
  consultationFee?: number | null;
  /** @nullable */
  patientName?: string | null;
  /** @nullable */
  appointmentDate?: string | null;
  /** @nullable */
  notes?: string | null;
  status: string;
  createdAt: string;
}

export interface AppointmentInput {
  doctorNameAr: string;
  specialty?: string;
  clinicLocation?: string;
  whatsappNumber: string;
  consultationFee?: number;
  patientName?: string;
  appointmentDate?: string;
  notes?: string;
}

export interface CategoryCount {
  categorySlug: string;
  categoryNameAr: string;
  count: number;
}

export interface AdminStats {
  totalListings: number;
  pendingListings: number;
  approvedListings: number;
  rejectedListings: number;
  totalCategories: number;
  totalAppointments: number;
  listingsByCategory: CategoryCount[];
}

export type ListListingsParams = {
category?: string;
search?: string;
page?: number;
limit?: number;
};

export type SearchListingsParams = {
q: string;
category?: string;
};

export type AdminListListingsParams = {
status?: string;
category?: string;
};