# lib/api-zod/src/generated/types/listing.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Wardan Online Arabic Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { ListingStatus } from './listingStatus';

export interface Listing {
  id: number;
  titleAr: string;
  /** @nullable */
  descriptionAr?: string | null;
  categorySlug: string;
  /** @nullable */
  categoryNameAr?: string | null;
  /** @nullable */
  price: number | null;
  /** @nullable */
  priceUnit?: string | null;
  /** @nullable */
  location?: string | null;
  whatsappNumber: string;
  /** @nullable */
  sellerName?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  status: ListingStatus;
  featured?: boolean;
  createdAt: string;
}