# lib/api-zod/src/generated/types/listingInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Wardan Online Arabic Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface ListingInput {
  /** @minLength 1 */
  titleAr: string;
  descriptionAr?: string;
  categorySlug: string;
  price?: number;
  priceUnit?: string;
  location?: string;
  whatsappNumber: string;
  sellerName?: string;
  imageUrl?: string;
}