# lib/api-zod/src/generated/types/listingUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Wardan Online Arabic Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { ListingUpdateStatus } from './listingUpdateStatus';

export interface ListingUpdate {
  titleAr?: string;
  descriptionAr?: string;
  price?: number;
  priceUnit?: string;
  location?: string;
  whatsappNumber?: string;
  sellerName?: string;
  imageUrl?: string;
  status?: ListingUpdateStatus;
  featured?: boolean;
}